bug fixes, should be runnable.
I’ve been getting a weird issue where the socket is busy when trying to
run the website, If you ever get that issue (and you have a Mac, sorry
Windows user idk) then go to the README and run the last line of code
in your terminal, then kill the process that flask is running on, and
rerun the program. Talk to me if you have an issue, I’ve gotten good at
working through the stupid problem.